Refactor GH query types & consts to separate file
